Material Generator issue 230.
Fairly substantial overhaul of the generator logic, to compute the
multiplier based on the required fuel(s) to allow for partial-generation
when the upgrade has a capacity modifier. Additionally fixes a bug in
the fuel consumption, where the full amount of fuel was consumed even if
a smaller number of items were ultimately added to the pedestal (due to
it being at capacity).