Simplification refactor in preparation for updating Quarry and Pump.
Mainly includes simplifications of the Pump and Quarry code to fold
together some existing conditionals.
Changes some of the Pump functions to operate directly on the FluidStack
that would be returned from the target block, so that we don't have to
re-compute that (and have logic to figure that out) all over the place.
Updates passesMachineFilter to passesMachineFilterItems and adds a new
passesMachineFilterFluids to ItemUpgradeBase (to cut down on repeated
code).