Add Operate to Bedrock infusion for ItemUpgradePump.
Includes a pretty compelte re-write of the upgrade. It now respects the
defined work area regardless of size unless the new infusion is in place
(allowing for larger than 4 sized lava cauldron setups). It constructs
the list of blocks to iterate over taking into account the "height" of
blocks it operates on at once (so as to not revisit blocks it previously
would have pumped due to being at a higher y-level). It only operates
when it has space for *all* of the fluid it would pump in a given
action, meaning if you have enough block manipulation infusions you will
need storage augments on the pedestal.