Add a new Remote Inventory Targeting infusion, used in ExportUpgrade.
This infusion modifies the upgrade to instead use a Locations Work Card,
allowing up to 8 locations to be defined where exports will be
attempted. This works similarly to linking together pedestals in that
once transfer is successful (even if just for one of the types being
transferred but not the rest) then the export is considered completed.